+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Functionality to allow chunking of fastq outputs from ingress.
+
 ## [v5.0.4]
 ### Fixed
 - `xam_ingress` failing to stage S3 BAM indexes.